requirements.txt
imutils==0.5.3
joblib==0.13.2
numpy
Pillow==6.0.0
scikit-image
scikit-learn

